refactor: 폴더별 index.js 생성
