+++ src/domain/index.js
@@ -1,0 +1,5 @@
+const Car = require('./Car');
+const MoveDecider = require('./MoveDecider');
+const RacingCarGame = require('./RacingCarGame');
+
+module.exports = { Car, MoveDecider, RacingCarGame };
