+++ src/utils/index.js
@@ -1,0 +1,5 @@
+const Console = require('./console');
+const pickNumberInRange = require('./pickNumberInRange');
+const Validator = require('./validator');
+
+module.exports = { Console, pickNumberInRange, ...Validator };
