+++ src/view/index.js
@@ -1,0 +1,4 @@
+const InputView = require('./InputView');
+const OutputView = require('./OutputView');
+
+module.exports = { InputView, OutputView };
